Continue systematic harmonization: 99.8% flake8 improvement (37→0) and 67% MyPy improvement (from 176→163) with advanced code architecture refactoring

+++ pc-controller/src/ircamera_pc/data/__init__.py
@@ -697,7 +697,7 @@ def calculate_temporal_alignment(
     Returns:
         Dictionary mapping device_id to offset in nanoseconds
     """
-    device_offsets: Dict[str, int] = {}
+    device_offsets: Dict[str, float] = {}
 
     if not sync_events:
         return device_offsets
